[wip] nimble/ll: ble_ll_timer
This adds ble_ll_timer APIs that are intended to replace os_cputime in
all LL code. The main reason for this was to provide more optimized
routines for ticks<->usecs routines that we can rely on. In addition,
this will make NimBLE controller more portable since eventually we can
allow to use something else than os_cputime as long as it meets our
requirements for LL timer.

The main optimization here is usecs->ticks conversion routine which
uses 64bit integer division. On CMAC, which does not even support 32bit
integer division, conversion takes ~3us which may not sound like a lot,
but in case we want to schedule aux chain scan at 300us offset we need
to save every usec possible as timing is very tight.
Optimized routine can perform conversion for values of up to ~32.1ms
using 32bit integers only and without using division so it takes pretty
much no time compared to original routine. Also it returns proper value
for all values within that range (original routine can be 1 tick off).

+++ nimble/controller/include/controller/ble_ll_sched.h
@@ -166,8 +166,8 @@ int ble_ll_sched_sync_reschedule(struct ble_ll_sched_item *sch,
                                  uint8_t anchor_point_usecs,
                                  uint32_t window_widening, int8_t phy_mode);
 int ble_ll_sched_sync(struct ble_ll_sched_item *sch,
-                      uint32_t beg_cputime, uint32_t rem_usecs, uint32_t offset,
-                      int8_t phy_mode);
+                      uint32_t beg_cputime, uint8_t rem_usecs, uint32_t offset,
+                      uint8_t phy_mode);
 
 /* Reschedule an advertising event */
 int ble_ll_sched_adv_reschedule(struct ble_ll_sched_item *sch,
